Move test to appropriate file
